+++ src/providers/eigenlayer/facets/EigenCoverageProviderFacet.sol
@@ -362,9 +362,7 @@ contract EigenCoverageProviderFacet is EigenCoverageStorage, ICoverageProvider {
 
         // Swap the slashed strategy asset to the coverage agent's asset
         IAssetPriceOracleAndSwapper(address(this)).
-            // casting is safe because we know the amount can not be larger than a uint128
-            // forge-lint: disable-next-line(unsafe-typecast)
-            swapForOutput(uint128(amount), ICoverageAgent(_position.coverageAgent).asset(), _position.asset);
+            swapForOutput(amount, ICoverageAgent(_position.coverageAgent).asset(), _position.asset);
 
         // Transfer swapped tokens to coverage agent
         bool success = IERC20(ICoverageAgent(_position.coverageAgent).asset()).transfer(_position.coverageAgent, amount);
